requirements.txt
# Wyckoff-Elliott Discord Trading Signals Bot

# Core dependencies
discord.py>=2.3.2
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3

# Binance interaction
python-binance>=1.0.19
websockets>=12.0

# Supabase
supabase>=2.0.0
postgrest>=0.13.0
realtime>=1.0.0

# Data processing & analysis
pandas>=2.1.0
numpy>=1.24.0
pandas-ta>=0.3.14b

# Technical indicators (additional)
ta-lib>=0.4.28

# Backtesting & visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Optional: ML for false-positive filtering
scikit-learn>=1.3.0

# Optional: Redis cache
redis>=5.0.0
hiredis>=2.2.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
isort>=5.12.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
pydantic>=2.4.0
